// custom-plugins/plugin-chart-pivot-tableRx1/src/types.ts
export type RgbColor = {
  r: number;
  g: number;
  b: number;
  a?: number;
};

export type PivotTableTheme = {
  headerBg?: string;
  headerText?: string;
  rowHeaderBg?: string;
  rowHeaderText?: string;
  totalBg?: string;
  totalText?: string;
  subtotalBg?: string;
  subtotalText?: string;
  cellBg?: string;
  cellText?: string;
  borderColor?: string;
};

export type FilterType = Record<string, DataRecordValue>;
export type SelectedFiltersType = Record<string, DataRecordValue[]>;

export type DateFormatter =
  | TimeFormatter
  | NumberFormatter
  | ((value: DataRecordValue) => string);

export type FormulaMetric = {
  label: string;
  expression: string;
  d3format?: string;
  hidden?: boolean;
};